X0.m.7 Separated spike mode's display range from the scan range.  (For adjusting sampling rate) Added X axis ticks every 10 channels to identify channel numbers. The spike mode display now automatically prints the selected range rather than hard coded text.
I'm zeroing in on my disturbance.  It's between channels 110 - 140 and it sweeps from channel 140 to 110 over a 5 second period roughly every 10 seconds.  (I can see it much better if I have the scan set to 128 for some reason.  scanning a smaller region seemed to hide the disturbance. Maybe it's near a harmonic or something.)